+++ projects/swimlane/ngx-ui/src/lib/components/list/list-header/list-header.component.ts
@@ -1,5 +1,8 @@
 import { Component, ContentChild, Input, TemplateRef, ViewChild, ViewEncapsulation } from '@angular/core';
 import { ListHeaderTemplateDirective } from './list-header-template.directive';
+import { ListSortComparator } from '../list-sort.utils';
+import { ListHeaderSortType } from '../models/list-header-sort-type.type';
+import { ListSortDirection } from '../models/list-sort-direction.type';
 
 @Component({
   selector: 'ngx-list-header',
@@ -14,8 +17,21 @@ import { ListHeaderTemplateDirective } from './list-header-template.directive';
 export class ListHeaderComponent {
   @ViewChild('template', { static: true }) template: TemplateRef<any>;
 
-  @Input() header: any;
+  @Input() header: ListHeaderComponent;
+  @Input() sortable = false;
+  @Input() prop?: string;
+  @Input() type?: ListHeaderSortType;
+  @Input() comparator?: ListSortComparator;
+  @Input() sortDir?: ListSortDirection;
+  @Input() onHeaderSort?: (sourceHeader: ListHeaderComponent) => void;
 
   @ContentChild(ListHeaderTemplateDirective, { read: TemplateRef, static: true })
   headerTemplate: TemplateRef<ListHeaderTemplateDirective>;
+
+  onSortClick(): void {
+    const sourceHeader = this.header ?? this;
+    if (sourceHeader.sortable) {
+      this.onHeaderSort?.(sourceHeader);
+    }
+  }
 }