+++ projects/swimlane/ngx-ui/src/lib/components/list/models/list-sort-prop-dir.ts
@@ -1,0 +1,6 @@
+import { ListSortDirection } from './list-sort-direction.type';
+
+export interface ListSortPropDir {
+  prop: string;
+  dir: ListSortDirection;
+}